* rename TransactionBuilder to TransactionSerializer * add TransactionBuilder with candidate builder interface
