+++ CardanoSharp.Wallet/Models/Transactions/TransactionBody/Metadata.cs
@@ -1,0 +1,6 @@
+namespace CardanoSharp.Wallet.Models.Transactions
+{
+    public class Metadata {
+
+    }
+}
